Add anchor nodes
