+++ src/Token/Anchor/Start.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace BackEndTea\Regexer\Token\Position;
+namespace BackEndTea\Regexer\Token\Anchor;
 
 use BackEndTea\Regexer\Token;
 
-final class End extends Token
+final class Start extends Token
 {
     public static function create(): self
     {
-        return new self('$');
+        return new self('^');
     }
 }
